+++ packages/eslint-config-expo/utils/expo.js
@@ -1,12 +1,19 @@
-module.exports = {
-  plugins: ['expo'],
-  ignorePatterns: [
+const expo = require('eslint-plugin-expo');
+
+module.exports = [
+  {
     // JS files can end up in build intermediates, eg:
     // android/app/build/intermediates/assets/debug/EXDevMenuApp.android.js
-    'android/app/build',
-  ],
-  rules: {
-    'expo/no-env-var-destructuring': ['error'],
-    'expo/no-dynamic-env-var': ['error'],
+    ignores: ['android/app/build'],
   },
-};
+  {
+    plugins: {
+      expo,
+    },
+
+    rules: {
+      'expo/no-env-var-destructuring': ['error'],
+      'expo/no-dynamic-env-var': ['error'],
+    },
+  },
+];